+++ spec/javascripts/app/views/aspects_list_view_spec.js
@@ -23,12 +23,12 @@ describe("app.views.AspectsList", function(){
       expect(this.view.$('.selected + a.selectable').text()).toMatch('Work');
     });
 
-    it('should show all the aspects', function(){
-      var aspect_selectors = this.view.$('.entypo.check + a.selectable');
-      expect(aspect_selectors.length).toBe(3);
-      expect(aspect_selectors[0].text).toMatch('Work');
-      expect(aspect_selectors[1].text).toMatch('Friends');
-      expect(aspect_selectors[2].text).toMatch('Acquaintances');
+    it("should show all the aspects", function(){
+      var aspectSelectors = this.view.$(".entypo-check + a.selectable");
+      expect(aspectSelectors.length).toBe(3);
+      expect(aspectSelectors[0].text).toMatch("Work");
+      expect(aspectSelectors[1].text).toMatch("Friends");
+      expect(aspectSelectors[2].text).toMatch("Acquaintances");
     });
 
     it('should show \'Select all\' link', function(){